requirements.txt
opencv-python~=4.5.1.48
numpy>=1.21
tqdm~=4.59.0
pillow~=9.0.0
visualdl~=2.1.1
scikit-image~=0.18.1

mxnet~=1.7.0.post2
jinja2~=3.1.4
pyyaml~=6.0.1
requests~=2.31.0
flask~=2.2.5
werkzeug~=2.2.3
gevent~=22.10.2
